backend/requirements.txt
alembic==1.16.5
beautifulsoup4==4.14.2
blinker==1.9.0
Brotli==1.1.0
cachelib==0.13.0
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cssselect==1.3.0
Django==4.2.15
djangorestframework==3.15.2
et_xmlfile==2.0.0
Flask==3.1.2
Flask-Caching==2.3.1
Flask-Compress==1.18
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Limiter==3.10.1
Flask-Talisman==1.1.0
Flask-WTF==1.2.1
fuzzywuzzy==0.18.0
greenlet==3.2.4
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Levenshtein==0.27.1
lxml==6.0.2
lxml_html_clean==0.4.3
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
playwright==1.55.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-Levenshtein==0.27.1
pytz==2025.2
pyzstd==0.18.0
RapidFuzz==3.14.1
readability-lxml==0.8.4.1
redis==5.1.0
requests==2.32.5
rq==1.16.2
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
xlrd==2.0.2
gunicorn==23.0.0
